requirements.txt
pygame==2.6.0
pip~=23.2.1
PyYAML~=6.0.2